backend/requirements.txt
flask==3.1.*
flask-cors==5.*
flask-caching==2.*
PyGithub==2.*
requests>=2.31
python-dotenv==1.*
gunicorn==23.*
